requirements.txt
fastapi
uvicorn
pydantic
httpx
matplotlib
plotly
kaleido
pytest
python-dotenv
databricks-sql-connector
